+++ src/main/java/cat/udl/eps/softarch/fll/controller/EditionCompetitionTableController.java
@@ -11,7 +11,7 @@ import org.springframework.web.bind.annotation.RequestMapping;
 import org.springframework.web.bind.annotation.RestController;
 import cat.udl.eps.softarch.fll.controller.dto.EditionCompetitionTableResponse;
 import cat.udl.eps.softarch.fll.exception.EditionCompetitionTableNotFoundException;
-import cat.udl.eps.softarch.fll.service.EditionCompetitionTableService;
+import cat.udl.eps.softarch.fll.service.edition.EditionCompetitionTableService;
 import lombok.RequiredArgsConstructor;
 
 @RestController
