+++ src/main/java/cat/udl/eps/softarch/fll/controller/MatchController.java
@@ -9,7 +9,7 @@ import org.springframework.web.bind.annotation.RequestMapping;
 import org.springframework.web.bind.annotation.RestController;
 import cat.udl.eps.softarch.fll.controller.dto.MatchTableAssignmentRequest;
 import cat.udl.eps.softarch.fll.controller.dto.MatchTableAssignmentResponse;
-import cat.udl.eps.softarch.fll.service.MatchTableAssignmentService;
+import cat.udl.eps.softarch.fll.service.match.MatchTableAssignmentService;
 import jakarta.validation.Valid;
 
 @RestController
