+++ src/main/java/cat/udl/eps/softarch/fll/controller/MatchScoreRegistrationController.java
@@ -10,7 +10,7 @@ import org.springframework.web.bind.annotation.RequestBody;
 import org.springframework.web.bind.annotation.RequestMapping;
 import org.springframework.web.bind.annotation.RestController;
 import cat.udl.eps.softarch.fll.controller.dto.ApiErrorResponse;
-import cat.udl.eps.softarch.fll.service.MatchScoreRegistrationService;
+import cat.udl.eps.softarch.fll.service.match.MatchScoreRegistrationService;
 import io.swagger.v3.oas.annotations.Operation;
 import io.swagger.v3.oas.annotations.tags.Tag;
