+++ src/main/java/cat/udl/eps/softarch/fll/handler/MatchEventHandler.java
@@ -5,7 +5,7 @@ import org.springframework.data.rest.core.annotation.HandleBeforeSave;
 import org.springframework.data.rest.core.annotation.RepositoryEventHandler;
 import org.springframework.stereotype.Component;
 import cat.udl.eps.softarch.fll.domain.Match;
-import cat.udl.eps.softarch.fll.service.MatchScheduleValidationService;
+import cat.udl.eps.softarch.fll.service.match.MatchScheduleValidationService;
 
 @Component
 @RepositoryEventHandler
