+++ src/main/java/cat/udl/eps/softarch/fll/service/coach/CoachService.java
@@ -1,4 +1,4 @@
-package cat.udl.eps.softarch.fll.service;
+package cat.udl.eps.softarch.fll.service.coach;
 
 import org.springframework.stereotype.Service;
 import org.springframework.transaction.annotation.Transactional;
