+++ src/main/java/cat/udl/eps/softarch/fll/service/leaderboard/LeaderboardService.java
@@ -1,4 +1,4 @@
-package cat.udl.eps.softarch.fll.service;
+package cat.udl.eps.softarch.fll.service.leaderboard;
 
 import java.util.List;
 import org.springframework.data.domain.Page;
