+++ src/main/java/cat/udl/eps/softarch/fll/service/match/MatchAssignmentService.java
@@ -1,4 +1,4 @@
-package cat.udl.eps.softarch.fll.service;
+package cat.udl.eps.softarch.fll.service.match;
 
 import java.util.ArrayList;
 import java.util.Comparator;
