+++ src/main/java/cat/udl/eps/softarch/fll/service/match/MatchSearchService.java
@@ -1,4 +1,4 @@
-package cat.udl.eps.softarch.fll.service;
+package cat.udl.eps.softarch.fll.service.match;
 
 import java.time.LocalTime;
 import java.util.List;
