+++ src/main/java/cat/udl/eps/softarch/fll/service/ranking/BestRoundRankingCalculator.java
@@ -1,23 +1,21 @@
-package cat.udl.eps.softarch.fll.service;
+package cat.udl.eps.softarch.fll.service.ranking;
 
 import cat.udl.eps.softarch.fll.domain.MatchResult;
 import cat.udl.eps.softarch.fll.domain.Team;
 import java.util.HashMap;
 import java.util.List;
 import java.util.Map;
-import org.springframework.context.annotation.Primary;
 import org.springframework.stereotype.Component;
 
-@Primary
 @Component
-public class TotalScoreRankingCalculator extends RankingCalculator {
+public class BestRoundRankingCalculator extends RankingCalculator {
 
 	@Override
 	protected Map<Team, Integer> computeScorePerTeam(List<MatchResult> results) {
-		Map<Team, Integer> totalScore = new HashMap<>();
+		Map<Team, Integer> bestScore = new HashMap<>();
 		for (MatchResult result : results) {
-			totalScore.merge(result.getTeam(), result.getScore(), Integer::sum);
+			bestScore.merge(result.getTeam(), result.getScore(), Math::max);
 		}
-		return totalScore;
+		return bestScore;
 	}
 }
