+++ src/main/java/cat/udl/eps/softarch/fll/service/ranking/TotalScoreRankingCalculator.java
@@ -1,21 +1,23 @@
-package cat.udl.eps.softarch.fll.service;
+package cat.udl.eps.softarch.fll.service.ranking;
 
 import cat.udl.eps.softarch.fll.domain.MatchResult;
 import cat.udl.eps.softarch.fll.domain.Team;
 import java.util.HashMap;
 import java.util.List;
 import java.util.Map;
+import org.springframework.context.annotation.Primary;
 import org.springframework.stereotype.Component;
 
+@Primary
 @Component
-public class BestRoundRankingCalculator extends RankingCalculator {
+public class TotalScoreRankingCalculator extends RankingCalculator {
 
 	@Override
 	protected Map<Team, Integer> computeScorePerTeam(List<MatchResult> results) {
-		Map<Team, Integer> bestScore = new HashMap<>();
+		Map<Team, Integer> totalScore = new HashMap<>();
 		for (MatchResult result : results) {
-			bestScore.merge(result.getTeam(), result.getScore(), Math::max);
+			totalScore.merge(result.getTeam(), result.getScore(), Integer::sum);
 		}
-		return bestScore;
+		return totalScore;
 	}
 }
