+++ src/test/java/cat/udl/eps/softarch/fll/edition/EditionCompetitionTableControllerTest.java
@@ -15,7 +15,7 @@ import org.springframework.test.web.servlet.setup.MockMvcBuilders;
 import cat.udl.eps.softarch.fll.controller.dto.EditionCompetitionTableResponse;
 import cat.udl.eps.softarch.fll.controller.dto.EditionTableMatchResponse;
 import cat.udl.eps.softarch.fll.exception.EditionCompetitionTableNotFoundException;
-import cat.udl.eps.softarch.fll.service.EditionCompetitionTableService;
+import cat.udl.eps.softarch.fll.service.edition.EditionCompetitionTableService;
 
 class EditionCompetitionTableControllerTest {
