+++ src/test/java/cat/udl/eps/softarch/fll/edition/EditionLifecycleControllerTest.java
@@ -16,7 +16,7 @@ import cat.udl.eps.softarch.fll.exception.DomainValidationExceptionHandler;
 import cat.udl.eps.softarch.fll.domain.EditionState;
 import cat.udl.eps.softarch.fll.exception.EditionLifecycleException;
 import cat.udl.eps.softarch.fll.handler.EditionLifecycleExceptionHandler;
-import cat.udl.eps.softarch.fll.service.EditionLifecycleService;
+import cat.udl.eps.softarch.fll.service.edition.EditionLifecycleService;
 
 class EditionLifecycleControllerTest {
