+++ src/test/java/cat/udl/eps/softarch/fll/leaderboard/LeaderboardServiceTest.java
@@ -9,7 +9,7 @@ import static org.mockito.Mockito.verify;
 import static org.mockito.Mockito.when;
 import java.util.List;
 
-import cat.udl.eps.softarch.fll.service.LeaderboardService;
+import cat.udl.eps.softarch.fll.service.leaderboard.LeaderboardService;
 import org.junit.jupiter.api.Test;
 import org.junit.jupiter.api.extension.ExtendWith;
 import org.mockito.InjectMocks;
