+++ src/test/java/cat/udl/eps/softarch/fll/match/MatchControllerTest.java
@@ -17,7 +17,7 @@ import org.springframework.validation.beanvalidation.LocalValidatorFactoryBean;
 import org.springframework.web.server.ResponseStatusException;
 import cat.udl.eps.softarch.fll.controller.dto.MatchTableAssignmentResponse;
 import cat.udl.eps.softarch.fll.exception.GlobalExceptionHandler;
-import cat.udl.eps.softarch.fll.service.MatchTableAssignmentService;
+import cat.udl.eps.softarch.fll.service.match.MatchTableAssignmentService;
 
 class MatchControllerTest {
