+++ src/test/java/cat/udl/eps/softarch/fll/match/MatchScoreRegistrationServiceTest.java
@@ -5,8 +5,8 @@ import cat.udl.eps.softarch.fll.domain.MatchResult;
 import cat.udl.eps.softarch.fll.domain.Team;
 import cat.udl.eps.softarch.fll.repository.match.MatchRepository;
 import cat.udl.eps.softarch.fll.repository.match.MatchResultRepository;
-import cat.udl.eps.softarch.fll.service.MatchScoreRegistrationService;
-import cat.udl.eps.softarch.fll.service.RankingService;
+import cat.udl.eps.softarch.fll.service.match.MatchScoreRegistrationService;
+import cat.udl.eps.softarch.fll.service.ranking.RankingService;
 import org.junit.jupiter.api.BeforeEach;
 import org.junit.jupiter.api.Test;
 import org.junit.jupiter.api.extension.ExtendWith;
